Removing XML from the Zope 3 core into its own package.
Things are moving mostly to a new zopexml top-level package that is
the start of the new zope3 XML project. This will be checked in into
the Infrae CVS initially (http://cvs.infrae.com/z3/zopexml) for the
time being, but we should look for a better place.

The code to make automatic interface assignments to XMLDocuments
based on XML schema references in the XML document has been removed
for the time being, as it was not working anyway after recent refactoring.
We may take the simpler approach in using normal Zope 3 interfaces to
stand in for XML schema, Relax NG schema and DTDs. This removes the need
to mess with the ZCML innards and the need for a ZCML directive.